requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
youtube-transcript-api==1.1.1
requests==2.31.0
pydantic==2.5.0
google-cloud-logging==3.8.0
